Rename to hidesig and adds debug info
