+++ lua/hidesig/highlight_utils.lua
