+++ lua/hidesig/internal.lua
@@ -1,5 +1,5 @@
-local highlight_utils = require("nvim-hidesig.highlight_utils")
-local util = require("nvim-hidesig.util")
+local highlight_utils = require("hidesig.highlight_utils")
+local util = require("hidesig.util")
 
 local hidesig = {}
 
@@ -11,10 +11,13 @@ function hidesig.traverseNode(bufnr, node)
   local highlightGroup = highlight_utils.getHightlightGroupForRange(bufnr, startLine, startCol)
   local color = highlight_utils.getHighlightGroupColor(highlightGroup)
   local darkenValue = hidesig.configs.opacity or 0.75
+  local newHighlightGroup = string.format("%sDimmed", highlightGroup)
+  vim.notify("[traverse node] bufnr", bufnr)
+  vim.notify("[traverse node] node type", node:type())
 
   vim.api.nvim_set_hl(
     hidesig.ns,
-    string.format("%sDimmed", highlightGroup),
+    newHighlightGroup,
     {
       fg = util.darken(color, darkenValue),
       undercurl = false,
@@ -25,17 +28,19 @@ function hidesig.traverseNode(bufnr, node)
   vim.api.nvim_buf_add_highlight(
     bufnr,
     hidesig.ns,
-    string.format("%sDimmed", highlightGroup),
+    newHighlightGroup,
     startLine,
     startCol,
     endCol
   )
 
-  if node:child_count() == 0 then
+  if node:child_count() < 1 then
+    vim.notify("node has no children")
     return
   else
+    vim.notify("node children")
     for childNode in node:iter_children() do
-      traverseNode(childNode)
+      hidesig.traverseNode(bufnr, childNode)
     end
   end
 end
@@ -44,6 +49,7 @@ end
 ---@param configs table configuration for hidesig
 -- @param configs.opacity float value from 0.0 to 1.0
 function hidesig.setup(configs)
+  vim.notify("[hidesig] setting up")
   hidesig.ns = vim.api.nvim_create_namespace("hidesig_ns")
   hidesig.configs = configs or {}
 end
@@ -75,13 +81,15 @@ function hidesig.perform(bufnr, lang)
   -- TODO: May need to clear highlight before adding running highlight. Otherwise commented code may retain the highlight
   -- reference https://github.dev/p00f/nvim-ts-rainbow/blob/master/lua/rainbow/internal.lua:51
   vim.api.nvim_buf_clear_namespace(bufnr, hidesig.ns, 0, -1)
+  local count = 0
   for _, captures in parsedQuery:iter_matches(bufferRoot, bufnr) do
     local sigBlock = captures[2] -- capture @sig_def
-
-    for rootChildNode in sigBlock:iter_children() do
-      hidesig.traverseNode(bufnr, rootChildNode)
-    end
+    count = count + 1
+    -- for rootChildNode in sigBlock:iter_children() do
+      -- hidesig.traverseNode(bufnr, rootChildNode)
+    -- end
   end
+  vim.notify(count)
 end
 
 return hidesig
