+++ lua/hidesig/util.lua
