--- plugin/nvim-hidesig.vim
@@ -1,1 +1,0 @@
-lua require("nvim-hidesig").init()
