+++ queries/ruby/sig_def.scm
@@ -1,6 +1,6 @@
 (
   call
     method: (identifier) @sig_keyword
-    block:  (_)
+    block:  [(block) (do_block)]
   (#eq? @sig_keyword "sig")
 ) @sig_def
